requirements.txt
fastapi[all]
pymongo
python-dotenv
uvicorn
Jinja2